refactor(contracts): extract shared api contracts package
