+++ packages/contracts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@anilog/api",
+  "name": "@anilog/contracts",
   "version": "0.1.0",
   "type": "module",
   "main": "./src/index.ts",
@@ -14,13 +14,10 @@
   },
   "dependencies": {
     "@anilog/config": "workspace:*",
-    "@anilog/contracts": "workspace:*",
-    "@anilog/db": "workspace:*",
-    "drizzle-orm": "^0.45.1",
     "elysia": "^1.4.26"
   },
   "devDependencies": {
-    "@types/bun": "latest",
-    "typescript": "^5.6.3"
+    "@types/bun": "catalog:",
+    "typescript": "catalog:"
   }
 }
