+++ packages/contracts/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@anilog/config/tsconfig.base.json",
+  "compilerOptions": {
+    "declaration": true,
+    "declarationMap": true,
+    "sourceMap": true,
+    "outDir": "dist",
+    "composite": true
+  }
+}
